Make it internal and rename it (#159)
This function seems useful for developer but users should likely not
interact with it directy. It should be called not by a user but by
another function that wants to assert that the data has no missing value.

Thus the function should not be exported and also not tested directy but
instead through the higher level, exported function that depends on it.