In tests avoid indirection
Whenever possible, it's best for each test to be fully self-contained.
When a test fails it's easier to understand why when you can see all
the context you need in the test itself.

I may have violated this principle myself. For example, by writing
helper functions that create toy data. While that sometimes help
make tests easier to write, it makes them harder to read -- so bad
on me.